Fix type-tracking steps (#50)
* Fix additional taint/type-tracking steps.

The propagation-graph implementation was relying on an internal API which recently changed. Fortunately, there is no an official API to achieve what we want to do, so we can just use that.

* Fix propagation-graph tests.

The standard library's model of `Map` has improved to the point where it was giving us flow in our getter/setter tests. We actually want to test our own heuristics for getter/setter detection, though, so this commit renames `Map` to `FancyMap`, and `get`/`set` to `load`/`store`.

* Update expected test output.